Go: Move project analysis code to separate file
